[#8] Refactor ProtonBasedDownstreamSender into subclasses.
Having separate types for the downstream senders makes the qualifiers obsolete
and allows to leverage conditional bean creation for multiple implementations
by retrieving the beans from Spring's application context.
